+++ src/server/game/Maps/MapManager.h
@@ -117,7 +117,7 @@ class MapManager
             {
                 float mod = o *-1;
                 mod = fmod(mod, 2.0f * static_cast<float>(M_PI));
-                mod = -mod + 2.0f * M_PI;
+                mod = -mod + 2.0f * static_cast<float>(M_PI);
                 return mod;
             }
             return fmod(o, 2.0f * static_cast<float>(M_PI));
